+++ src/spec/jsx-fn-interoperability.spec.tsx
@@ -1,5 +1,5 @@
-import {jsx, h} from '..';
-import {StaticArg, StaticChild} from '../types';
+import {jsx, h, getElement} from '..';
+import {Fusion, StaticArg, StaticChild} from '../types';
 
 describe('children wrapper', () => {
   const IdiomaticJsx = ({children}: {children?: StaticChild}) => {
@@ -15,62 +15,66 @@ describe('children wrapper', () => {
     );
   };
 
-  const IdiomaticFn = (...children: StaticChild[]) => {
+  const IdiomaticFn = (...children: any[]) => {
     return h('idiomatic-fn', ...children);
   };
 
-  const CompatibleFn = (...args: StaticArg[]) => {
+  const CompatibleFn = (...args: any[]) => {
     return h('compatible-fn', ...args);
   };
 
-  test.each(
-    // prettier-ignore
+  test.each<[Fusion, string]>([
+    // idiomatic jsx component
     [
-      // idiomatic jsx component
-      [
-        (<jsx><IdiomaticJsx>AAA{111}</IdiomaticJsx></jsx>) as Element,
-        '<jsx><idiomatic-jsx>AAA111</idiomatic-jsx></jsx>',
-      ],
-      [
-        h('fn', IdiomaticJsx({children: ['AAA', 111]}) as Element),
-        '<fn><idiomatic-jsx>AAA111</idiomatic-jsx></fn>',
-        // ! children defined in props
-      ],
-
-      // compatible jsx component
-      [
-        (<jsx><CompatibleJsx>AAA{111}</CompatibleJsx></jsx>) as Element,
-        '<jsx><compatible-jsx>AAA111</compatible-jsx></jsx>',
-      ],
-      [
-        h('fn', CompatibleJsx('AAA', 111) as Element),
-        '<fn><compatible-jsx>AAA111</compatible-jsx></fn>',
-      ],
-
-      // idiomatic fn component
-      [
-        h('fn', IdiomaticFn('AAA', 111)),
-        '<fn><idiomatic-fn>AAA111</idiomatic-fn></fn>',
-      ],
-      [
-        // @ts-ignore
-        (<jsx><IdiomaticFn>AAA{111}</IdiomaticFn></jsx>) as Element,
-        '<jsx><idiomatic-fn>AAA111</idiomatic-fn></jsx>',
-        // ! children are not allowed in typescript
-        // it works bacause of {children} jsx support in props in initFn()
-      ],
-
-      // compatible fn component
-      [
-        h('fn', CompatibleFn('AAA', 111)),
-        '<fn><compatible-fn>AAA111</compatible-fn></fn>',
-      ],
-      [
-        (<jsx><CompatibleFn>AAA{111}</CompatibleFn></jsx>) as Element,
-        '<jsx><compatible-fn>AAA111</compatible-fn></jsx>',
-      ]
+      <jsx>
+        <IdiomaticJsx>AAA{111}</IdiomaticJsx>
+      </jsx>,
+      '<jsx><idiomatic-jsx>AAA111</idiomatic-jsx></jsx>',
     ],
-  )('%p %p', (element, html) => {
-    expect(element.outerHTML).toBe(html);
+    [
+      h('fn', IdiomaticJsx({children: ['AAA', 111]})),
+      '<fn><idiomatic-jsx>AAA111</idiomatic-jsx></fn>',
+      // ! children defined in props
+    ],
+
+    // compatible jsx component
+    [
+      <jsx>
+        <CompatibleJsx>AAA{111}</CompatibleJsx>
+      </jsx>,
+      '<jsx><compatible-jsx>AAA111</compatible-jsx></jsx>',
+    ],
+    [
+      h('fn', CompatibleJsx('AAA', 111)),
+      '<fn><compatible-jsx>AAA111</compatible-jsx></fn>',
+    ],
+
+    // idiomatic fn component
+    [
+      h('fn', IdiomaticFn('AAA', 111)),
+      '<fn><idiomatic-fn>AAA111</idiomatic-fn></fn>',
+    ],
+    [
+      <jsx>
+        <IdiomaticFn>AAA{111}</IdiomaticFn>
+      </jsx>,
+      '<jsx><idiomatic-fn>AAA111</idiomatic-fn></jsx>',
+      // ! children are not allowed in typescript
+      // it works bacause of {children} jsx support in props in initFn()
+    ],
+
+    // compatible fn component
+    [
+      h('fn', CompatibleFn('AAA', 111)),
+      '<fn><compatible-fn>AAA111</compatible-fn></fn>',
+    ],
+    [
+      <jsx>
+        <CompatibleFn>AAA{111}</CompatibleFn>
+      </jsx>,
+      '<jsx><compatible-fn>AAA111</compatible-fn></jsx>',
+    ],
+  ])('%p %p', (provided, expected) => {
+    expect(getElement(provided).outerHTML).toBe(expected);
   });
 });